requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.5
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.6
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.6
argparse==1.2.1
decorator==3.4.0
itsdangerous==0.24
mutagen==1.23
pbr==0.8.2
six==1.7.2
sqlalchemy-migrate==0.9.1
wsgiref==0.1.2
bcrypt==1.0.2
validate_email==1.2
pydns==2.3.6
